Revert "Expand auth ui tests"
