+++ sonar-project.properties
@@ -19,4 +19,4 @@ sonar.language=js
 # Limit analysis to JavaScript files in the specified sources
 sonar.inclusions=**/*.js
 # Exclude HTML, CSS, SCSS, test files, coverage directory, and setup file
-sonar.exclusions=**/*.html, **/*.css, **/*.scss, **/*.test.js, coverage/**, jest.setup.js, /test-db-utils/sample-data/**
+sonar.exclusions=**/*.html, **/*.css, **/*.scss, **/*.test.js, coverage/**, jest.setup.js, test-db-utils/sample-data/**
